Set the emulation mode for embedded .NET WebBrowser controls during application startup
